bs_detector/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-aws>=0.1.0

# MCP integration
mcp>=0.1.0
httpx>=0.25.0

# Evaluation
deepeval>=0.21.0
pytest>=7.0.0

# Notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
mermaid-py>=0.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
pandas>=2.0.0